Added ability to attach render to email
The option is also coded in to include the image in the email body, rather than as an attachment, but I've commented it out for now, as attachments were a little easier to manage than full HD renders in a message body. Still, the option to have it as a toggle switch exists.