Fixed image not saving bug
Renders do not save unless (write_still=True) is passed to the render operator. Since I wanted this to work out of the box and not require users to run a separate setup script every time I used an app handler to ensure that write_still=True is passed to the render command by default.
In testing a render_pre and render_init handler solution I kept encountering depsgraph errors, so I connected it to the depsgraph_update_pre handler, and that seems to resolve the issue.