Fixed unnamed render bug
You no longer have to ensure that you have named your blend file, or your render.
The blend file will automatically be called 'untitled.blend' if no other name is given.
The render will automatically be named after the time the file was opened, called - HourMinuteSecond.PNG - or whatever the file extension is

I was able to pass this info in through handlers, so you should be able to render and email and ALMOST unchanged scene.
Unfortunately, the catch for this is that you must make some kind of adjustment to you scene manually.
If you try to render on a completely untouched scene, it will still fail, however, if you make even a tiny change to the scene, it will work.
If you move, scale, rotate an object, change a material value, the scene's start or end frame, even if you immediately change it back, that will trigger the depsgraph to update and allow the addon to rename and save the render. Otherwise it gives the same "File not found" error.

I thought I'd be able to trigger an update by running one of those commands in a post_load handler, but it appears to only work when a user does it manually.